api/requirements.txt
fastapi==0.104.1
uvicorn==0.34.3
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0
motor==3.7.1
pymongo==4.13.2
python-jose==3.3.0
bcrypt==4.0.1
python-multipart
python-dotenv==1.1.1
PyYAML==6.0.2
numpy==1.26.4
pandas==2.3.0
scipy==1.15.3
pillow==11.2.1
